Move solver tests
