+++ tests/core/test_solve.py
@@ -1,20 +1,23 @@
 # Copyright (C) 2012 Anaconda, Inc
 # SPDX-License-Identifier: BSD-3-Clause
+from __future__ import annotations
+
 import copy
 import os
 import re
 import sys
 from importlib.metadata import version
 from pprint import pprint
+from typing import TYPE_CHECKING
 from unittest.mock import Mock, patch
 
 import pytest
 
 from conda.auxlib.ish import dals
-from conda.base.context import conda_tests_ctxt_mgmt_def_pol, context
+from conda.base.context import conda_tests_ctxt_mgmt_def_pol, context, reset_context
 from conda.common.compat import on_linux, on_mac, on_win
 from conda.common.io import env_var, env_vars
-from conda.core.solve import DepsModifier, UpdateModifier
+from conda.core.solve import DepsModifier, UpdateModifier, get_pinned_specs
 from conda.exceptions import (
     PackagesNotFoundError,
     ResolvePackageNotFound,
@@ -23,9 +26,9 @@ from conda.exceptions import (
 )
 from conda.models.channel import Channel
 from conda.models.enums import PackageType
+from conda.models.match_spec import MatchSpec
 from conda.models.records import PrefixRecord
 from conda.models.version import VersionOrder
-from conda.resolve import MatchSpec
 from conda.testing.helpers import (
     CHANNEL_DIR_V1,
     add_subdir,
@@ -39,6 +42,12 @@ from conda.testing.helpers import (
     get_solver_cuda,
     get_solver_must_unfreeze,
 )
+
+if TYPE_CHECKING:
+    from pytest import MonkeyPatch
+    from pytest_mock import MockerFixture
+
+    from conda.testing.fixtures import CondaCLIFixture, TmpEnvFixture
 
 pytestmark = pytest.mark.usefixtures("parametrized_solver_fixture")
 
@@ -3794,3 +3803,82 @@ def test_globstr_matchspec_non_compatible(tmpdir):
     with get_solver(tmpdir, specs) as solver:
         with pytest.raises((PackagesNotFoundError, ResolvePackageNotFound)):
             solver.solve_final_state()
+
+
+def test_pinned_specs_CONDA_PINNED_PACKAGES(monkeypatch: MonkeyPatch) -> None:
+    # Test pinned specs environment variable
+    specs = ("numpy 1.11", "python >3")
+
+    monkeypatch.setenv("CONDA_PINNED_PACKAGES", "&".join(specs))
+    reset_context()
+    assert context.pinned_packages == specs
+
+    pinned_specs = get_pinned_specs("/none")
+    assert pinned_specs != specs
+    assert pinned_specs == tuple(MatchSpec(spec, optional=True) for spec in specs)
+
+
+def test_pinned_specs_conda_meta_pinned(tmp_env: TmpEnvFixture):
+    # Test pinned specs conda environment file
+    specs = ("scipy ==0.14.2", "openjdk >=8")
+    with tmp_env() as prefix:
+        (prefix / "conda-meta" / "pinned").write_text("\n".join(specs) + "\n")
+
+        pinned_specs = get_pinned_specs(prefix)
+        assert pinned_specs != specs
+        assert pinned_specs == tuple(MatchSpec(spec, optional=True) for spec in specs)
+
+
+def test_pinned_specs_condarc(
+    tmp_env: TmpEnvFixture,
+    conda_cli: CondaCLIFixture,
+    mocker: MockerFixture,
+):
+    # Test pinned specs conda environment file
+    specs = ("requests ==2.13",)
+    with tmp_env() as prefix:
+        # mock active prefix
+        mocker.patch(
+            "conda.base.context.Context.active_prefix",
+            new_callable=mocker.PropertyMock,
+            return_value=str(prefix),
+        )
+
+        conda_cli("config", "--env", "--add", "pinned_packages", *specs)
+
+        pinned_specs = get_pinned_specs(prefix)
+        assert pinned_specs != specs
+        assert pinned_specs == tuple(MatchSpec(spec, optional=True) for spec in specs)
+
+
+def test_pinned_specs_all(
+    tmp_env: TmpEnvFixture,
+    conda_cli: CondaCLIFixture,
+    mocker: MockerFixture,
+    monkeypatch: MonkeyPatch,
+) -> None:
+    # Test pinned specs conda configuration and pinned specs conda environment file
+    specs1 = ("numpy 1.11", "python >3")
+    specs2 = ("scipy ==0.14.2", "openjdk >=8")
+    specs3 = ("requests=2.13",)
+    specs = (*specs1, *specs3, *specs2)
+
+    monkeypatch.setenv("CONDA_PINNED_PACKAGES", "&".join(specs1))
+    reset_context()
+    assert context.pinned_packages == specs1
+
+    with tmp_env() as prefix:
+        (prefix / "conda-meta" / "pinned").write_text("\n".join(specs2) + "\n")
+
+        # mock active prefix
+        mocker.patch(
+            "conda.base.context.Context.active_prefix",
+            new_callable=mocker.PropertyMock,
+            return_value=str(prefix),
+        )
+
+        conda_cli("config", "--env", "--add", "pinned_packages", *specs3)
+
+        pinned_specs = get_pinned_specs(prefix)
+        assert pinned_specs != specs
+        assert pinned_specs == tuple(MatchSpec(spec, optional=True) for spec in specs)
